* Now able to read BitTorrent data from the FyreDL XML history file!

+++ src/addurl.cpp
@@ -204,7 +204,7 @@ void AddURL::on_buttonBox_accepted()
                     gk_torrent_data.dlStatus = GekkoFyre::DownloadStatus::Stopped;
                     gk_torrent_data.down_dest = ui->file_dest_lineEdit->text().toStdString();
                     routines->writeTorrentItem(gk_torrent_data);
-                    emit sendDetails(gk_torrent_data.torrent_name, (double)(gk_torrent_data.num_pieces * gk_torrent_data.piece_length),
+                    emit sendDetails(gk_torrent_data.torrent_name, ((double)gk_torrent_data.num_pieces * (double)gk_torrent_data.piece_length),
                                      0, 0, 0, 0, GekkoFyre::DownloadStatus::Stopped, gk_torrent_data.magnet_uri, gk_torrent_data.down_dest,
                                      GekkoFyre::HashType::None, "", 0, true, "", GekkoFyre::DownloadType::Torrent);
                     return AddURL::done(QDialog::Accepted);
